+++ src/network_event/client_event.rs
@@ -10,7 +10,7 @@ use tap::TapFallible;
 
 use super::EventChannel;
 use crate::{
-    client::{map_entity::NetworkEntityMap, ClientState},
+    client::{ClientState, NetworkEntityMap},
     prelude::NetworkChannels,
     server::{ServerSet, ServerState, SERVER_ID},
     REPLICATION_CHANNEL_ID,